+++ responsive_data_grid/test/grid_api_behavior_test.dart
@@ -99,7 +99,7 @@ void main() {
     await tester.pump();
     await tester.pump(const Duration(milliseconds: 50));
 
-    await tester.tap(find.byIcon(Icons.menu).first);
+    await tester.tap(find.byIcon(Icons.more_vert).first);
     await tester.pump();
     await tester.pump(const Duration(milliseconds: 50));
 
@@ -134,7 +134,7 @@ void main() {
     await tester.pump();
     await tester.pump(const Duration(milliseconds: 50));
 
-    await tester.tap(find.byIcon(Icons.menu).first);
+    await tester.tap(find.byIcon(Icons.more_vert).first);
     await tester.pump();
     await tester.pump(const Duration(milliseconds: 50));